requirements.txt
pandas
transformers
torch
Pillow
opencv-python
pytesseract
ultralytics
pytest==7.4.3
opencv-python #cv2
